+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [1.2.0] - 2023-10-03
 - Updated to work in Elasticsearch v8.5.2
 
+### Changed
+- main, run_program: Removed gen_libs.get_inst call.
+
 
 ## [1.1.2] - 2022-12-22
 - Updated to work in Python 3 too